eventapis-13(#13) operationstore rename to emon/emon-lib
