+++ emon-lib/Dockerfile
@@ -1,0 +1,5 @@
+FROM frolvlad/alpine-oraclejdk8:slim
+MAINTAINER maaydin
+EXPOSE 8080
+ADD target/eventapis.jar /
+CMD java -jar eventapis.jar
