+++ emon-lib/docker-compose.yml
@@ -1,0 +1,4 @@
+version: '2'
+services:
+  challenge:
+    image: 'kloiasoft/eventapis:latest'
