+++ emon-lib/pom.xml
@@ -10,9 +10,9 @@
         <version>0.4.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>emon</artifactId>
+    <artifactId>emon-lib</artifactId>
 
-    <name>Eventapis Store and Monitoring</name>
+    <name>Eventapis Store and Monitoring Core</name>
 
     <properties>
         <spring.version>1.5.7.RELEASE</spring.version>
@@ -42,10 +42,14 @@
     <dependencies>
         <dependency>
             <groupId>com.kloia.eventapis</groupId>
-            <artifactId>emon-lib</artifactId>
+            <artifactId>java-api</artifactId>
             <version>${project.version}</version>
         </dependency>
-
+        <dependency>
+            <groupId>com.kloia.eventapis</groupId>
+            <artifactId>spring-integration</artifactId>
+            <version>${project.version}</version>
+        </dependency>
         <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
@@ -54,7 +58,7 @@
         <dependency>
             <groupId>com.hazelcast</groupId>
             <artifactId>hazelcast</artifactId>
-            <version>3.8.6</version>
+            <version>3.9.1</version>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
@@ -139,24 +143,12 @@
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter</artifactId>
-<!--            <exclusions>
-                <exclusion>
-                    <artifactId>logback-classic</artifactId>
-                    <groupId>ch.qos.logback</groupId>
-                </exclusion>
-            </exclusions>-->
         </dependency>
 
         <dependency>
             <groupId>org.springframework.kafka</groupId>
             <artifactId>spring-kafka</artifactId>
             <version>1.3.1.RELEASE</version>
- <!--           <exclusions>
-                <exclusion>
-                    <artifactId>commons-logging</artifactId>
-                    <groupId>commons-logging</groupId>
-                </exclusion>
-            </exclusions>-->
         </dependency>
 
         <dependency>
@@ -169,33 +161,7 @@
 
 
     <build>
-        <!--<finalName>eventapis</finalName>-->
-        <resources>
-            <resource>
-                <directory>src/main/resources</directory>
-            </resource>
-            <resource>
-                <directory>src/main/static</directory>
-                <filtering>true</filtering>
-            </resource>
-        </resources>
         <plugins>
-            <!--            <plugin>
-                            <groupId>org.apache.maven.plugins</groupId>
-                            <artifactId>maven-jar-plugin</artifactId>
-                         serviceDiscoveryConfigurationtion>
-                                <archive>
-                                    <manifest>
-                                        <mainClass>com.kloia.eventapis.Eventapis</mainClass>
-                                    </manifest>
-                                </archive>
-                          serviceDiscoveryConfigurationtion>
-                        </plugin>-->
-            <!--   <plugin>
-                   <groupId>org.springframework.boot</groupId>
-                   <artifactId>spring-boot-maven-plugin</artifactId>
-                   <version>${spring.version}</version>
-               </plugin>-->
         </plugins>
     </build>
