+++ emon-lib/src/main/java/com/kloia/eventapis/api/emon/domain/EventHandler.java
@@ -1,4 +1,4 @@
-package com.kloia.eventapis.api.store.domain;
+package com.kloia.eventapis.api.emon.domain;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonTypeName;
