+++ emon-lib/src/main/java/com/kloia/eventapis/api/emon/domain/NoneHandler.java
@@ -1,4 +1,4 @@
-package com.kloia.eventapis.api.store.domain;
+package com.kloia.eventapis.api.emon.domain;
 
 import com.fasterxml.jackson.annotation.JsonTypeName;
 import com.kloia.eventapis.pojos.Operation;
