+++ emon-lib/src/test/java/com/kloia/eventapis/pojos/OperationTest.java
@@ -1,0 +1,14 @@
+package com.kloia.eventapis.pojos;
+
+import org.junit.Test;
+
+/**
+ * Created by zeldalozdemir on 26/01/2017.
+ */
+public class OperationTest {
+    @Test
+    public void testReadWriteExternal() throws Exception {
+
+    }
+
+}
