+++ emon-vis/app/templates/topology.hbs
@@ -13,7 +13,7 @@
         <h3>Latest Operations</h3>
         {{#each operation.latestOperations.operations as |operation|}}
           <li>
-            {{#link-to "topology" operation.opId}}
+            {{#link-to "topology" operation.opId preventDefault=false}}
               {{operation.initiatorCommand}}@{{operation.initiatorService}}/{{operation.opId}}
             {{/link-to}}
           </li>
